Phase 4 완료: 클라우드 동기화 기능 구현
- 데이터 동기화 서비스 (SyncService)
- Supabase Realtime 실시간 동기화
- 오프라인 지원 (OfflineQueueService)
- 백업/복원 기능 (BackupRestoreService)
- connectivity_plus, path_provider, share_plus, file_picker 패키지 추가
- 전체 진행률 80% 달성

+++ lib/screens/goal_setting_screen.dart
@@ -299,10 +299,10 @@ class _GoalSettingScreenState extends ConsumerState<GoalSettingScreen> {
                   Container(
                     padding: const EdgeInsets.all(20),
                     decoration: BoxDecoration(
-                      color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)).withOpacity(0.1),
+                      color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())).withOpacity(0.1),
                       borderRadius: BorderRadius.circular(16),
                       border: Border.all(
-                        color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)),
+                        color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())),
                         width: 2,
                       ),
                     ),
@@ -313,7 +313,7 @@ class _GoalSettingScreenState extends ConsumerState<GoalSettingScreen> {
                           children: [
                             Icon(
                               Icons.flag_outlined,
-                              color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)),
+                              color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())),
                               size: 24,
                             ),
                             const SizedBox(width: 12),
@@ -321,7 +321,7 @@ class _GoalSettingScreenState extends ConsumerState<GoalSettingScreen> {
                               '목표 BMI',
                               style: TextStyle(
                                 fontSize: 16,
-                                color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)),
+                                color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())),
                                 fontWeight: FontWeight.w600,
                               ),
                             ),
@@ -330,7 +330,7 @@ class _GoalSettingScreenState extends ConsumerState<GoalSettingScreen> {
                               targetBMI.toStringAsFixed(1),
                               style: TextStyle(
                                 fontSize: 28,
-                                color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)),
+                                color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())),
                                 fontWeight: FontWeight.bold,
                               ),
                             ),
@@ -338,10 +338,10 @@ class _GoalSettingScreenState extends ConsumerState<GoalSettingScreen> {
                         ),
                         const SizedBox(height: 8),
                         Text(
-                          BMICalculator.getCategoryName(BMICalculator.getBMICategory(targetBMI)),
+                          BMICalculator.getCategoryName(BMICalculator.getBMICategory(targetBMI.toDouble())),
                           style: TextStyle(
                             fontSize: 16,
-                            color: _getBMIColor(BMICalculator.getBMICategory(targetBMI)),
+                            color: _getBMIColor(BMICalculator.getBMICategory(targetBMI.toDouble())),
                             fontWeight: FontWeight.w600,
                           ),
                         ),
